% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{note_synth}
\alias{note_synth}
\title{Synthesize a Single Note}
\usage{
note_synth(
  freq,
  signal = "tria",
  duration_sec = 1,
  velocity = 100,
  sample_rate_hz = 48000,
  bit_width = 24
)
}
\arguments{
\item{freq}{The frequency of the note in Hz}

\item{signal}{Signal type: "sine", "tria", "square" or "saw"}

\item{duration_sec}{Duration in seconds, default 1}

\item{velocity}{MIDI velocity (0-127), default 100}

\item{sample_rate_hz}{Sample rate, default 48000}

\item{bit_width}{Bit width (8, 16, 24, 32, 64), default 24}
}
\value{
A tuneR Wave object
}
\description{
Generates a Wave object for a single frequency
}
